Refactor main.rb to only shows menu
